Refactoring the LearnAbout test. Moving them out of the CDSTest class. Optimizing a couple of the tests.
