+++ test/src/org/labkey/test/tests/cds/CDSTest.java
@@ -1210,322 +1210,6 @@ public class CDSTest extends CDSReadOnlyTest
     }
 
     @Test
-    public void testLearnAboutStudies()
-    {
-        cds.viewLearnAboutPage("Studies");
-
-        List<String> studies = Arrays.asList(CDSHelper.STUDIES);
-        _asserts.verifyLearnAboutPage(studies);
-    }
-
-    @Test
-    public void clickOnLearnAboutStudyItem()
-    {
-        List<WebElement> returnedItems;
-        String[] itemParts;
-        final String XPATH_RESULTLIST = "//div[contains(@class, 'learnview')]//span//div//div[contains(@class, 'learnstudies')]//div[contains(@class, 'learncolumnheader')]/./following-sibling::div[contains(@class, 'detail-container')]";
-
-        cds.viewLearnAboutPage("Studies");
-        sleep(CDSHelper.CDS_WAIT_ANIMATION);
-        returnedItems = Locator.xpath(XPATH_RESULTLIST).findElements(getDriver());
-
-        int index = returnedItems.size()/2;
-
-        itemParts = returnedItems.get(index).getText().split("\n");
-        returnedItems.get(index).click();
-
-        log("Validating title is " + itemParts[0]);
-        shortWait().until(ExpectedConditions.visibilityOfElementLocated(Locator.xpath("//div[contains(@class, 'learnheader')]//div//span[text()='" + itemParts[0] + "']").toBy()));
-
-        log("Validating Study Type is: " + itemParts[1]);
-        assert(Locator.xpath("//table[contains(@class, 'learn-study-info')]//tbody//tr//td[contains(@class, 'item-value')][text()='" + itemParts[1] + "']").findElement(getDriver()).isDisplayed());
-
-        log("Validating return link works.");
-        click(Locator.xpath("//div[contains(@class, 'learn-up')]/span[contains(@class, 'breadcrumb')][text()='Studies / ']"));
-
-        shortWait().until(ExpectedConditions.visibilityOfElementLocated(Locator.xpath("//div[contains(@class, 'title')][text()='Learn about...']").toBy()));
-    }
-
-    @Test
-    public void testLearnAboutStudiesSearch()
-    {
-        String searchString;
-        List<WebElement> returnedItems;
-        String itemText;
-        String[] itemParts;
-        final String XPATH_TEXTBOX = "//table[contains(@class, 'learn-search-input')]//tbody//tr//td//input";
-        final String XPATH_RESULTLIST = "//div[contains(@class, 'learnview')]//span//div//div[contains(@class, 'learnstudies')]//div[contains(@class, 'learncolumnheader')]/./following-sibling::div[contains(@class, 'detail-wrapper')]";
-
-        cds.viewLearnAboutPage("Studies");
-
-        searchString = "HVTN"; // TODO Test data dependent.
-        log("Searching for '" + searchString + "'.");
-        this.setFormElement(Locator.xpath(XPATH_TEXTBOX), searchString);
-        sleep(CDSHelper.CDS_WAIT_ANIMATION);  // Same elements are reused between searched, this sleep prevents a "stale element" error.
-        returnedItems = Locator.xpath(XPATH_RESULTLIST).findElements(getDriver());
-        log("Size: " + returnedItems.size());
-        for (WebElement listItem : returnedItems)
-        {
-            itemText = listItem.getText();
-            itemParts = itemText.split("\n");
-            log("Looking at study " + itemParts[0]);
-            assert(itemText.toLowerCase().contains(searchString.toLowerCase()));
-        }
-
-        searchString = "(vCP1452)"; // TODO Test data dependent.
-        log("Searching for '" + searchString + "'.");
-        this.setFormElement(Locator.xpath(XPATH_TEXTBOX), searchString);
-        sleep(CDSHelper.CDS_WAIT_ANIMATION);
-        returnedItems = Locator.xpath(XPATH_RESULTLIST).findElements(getDriver());
-        for (WebElement listItem : returnedItems)
-        {
-            itemText = listItem.getText();
-            itemParts = itemText.split("\n");
-            log("Looking at study " + itemParts[0]);
-            assert(itemText.toLowerCase().contains(searchString.toLowerCase()));
-        }
-
-        searchString = "Phase IIB"; // TODO Test data dependent.
-        log("Searching for '" + searchString + "'.");
-        this.setFormElement(Locator.xpath(XPATH_TEXTBOX), searchString);
-        sleep(CDSHelper.CDS_WAIT_ANIMATION);
-        returnedItems = Locator.xpath(XPATH_RESULTLIST).findElements(getDriver());
-        for (WebElement listItem : returnedItems)
-        {
-            itemText = listItem.getText();
-            itemParts = itemText.split("\n");
-            log("Looking at study " + itemParts[0]);
-            assert(itemText.toLowerCase().contains(searchString.toLowerCase()));
-        }
-
-        searchString = "If this string ever appears something very odd happened.";
-        log("Searching for '" + searchString + "'.");
-        sleep(CDSHelper.CDS_WAIT_ANIMATION);
-        this.setFormElement(Locator.xpath(XPATH_TEXTBOX), searchString);
-        _asserts.verifyEmptyLearnAboutStudyPage();
-
-    }
-
-    @Test
-    public void verifyLearnAboutStudyDetails()
-    {
-        // TODO Test data dependent.
-        final String searchString = "QED 2";
-        final String grantAffiliation = "Nulla tellus. In sagittis dui vel nisl.";
-        final String dataAvailability = "iaculis diam erat fermentum justo nec";
-        final String firstContactName = "Juan Owens";
-        final String firstContactEmail = "jowens5@deviantart.com";
-        final String rationale = "Nullam molestie nibh in lectus. Pellentesque at nulla.";
-        final String XPATH_TEXTBOX = "//table[contains(@class, 'learn-search-input')]//tbody//tr//td//input";
-        final String XPATH_RESULTLIST = "//div[contains(@class, 'detail-wrapper')]";
-
-        cds.viewLearnAboutPage("Studies");
-        log("Searching for '" + searchString + "'.");
-        this.setFormElement(Locator.xpath(XPATH_TEXTBOX), searchString);
-        sleep(CDSHelper.CDS_WAIT_ANIMATION);
-
-        log("Verifying data availability on summary page.");
-        assert(Locator.xpath("//div[contains(@class, 'data-availability-text')][text()='" + dataAvailability + "']").findElement(getDriver()).isDisplayed());
-
-        log("Start verifying study detail page.");
-        List<WebElement> returnedItems  = Locator.xpath(XPATH_RESULTLIST).findElements(getDriver());
-        returnedItems.get(0).click();
-        sleep(CDSHelper.CDS_WAIT_ANIMATION);
-
-        log("Verifying study information.");
-        List<String> fields = Arrays.asList(CDSHelper.LEARN_ABOUT_QED2_INFO_FIELDS);
-        fields.stream().forEach((field) ->  {
-            assertTextPresent(field);
-        });
-        assertTextPresent(grantAffiliation);
-
-        log("Verifying contact information.");
-        fields = Arrays.asList(CDSHelper.LEARN_ABOUT_QED2_CONTACT_FIELDS);
-        fields.stream().forEach((field) ->  {
-            assertTextPresent(field);
-        });
-
-        assertElementPresent(Locator.xpath("//a[contains(@href, 'mailto:" + firstContactEmail + "')][text()='" + firstContactName + "']"));
-
-        log("Verifying description section.");
-        fields = Arrays.asList(CDSHelper.LEARN_ABOUT_QED2_DESCRIPTION_FIELDS);
-        fields.stream().forEach((field) ->  {
-            assertTextPresent(field);
-        });
-        assertTextPresent(rationale);
-        assertElementPresent(Locator.xpath("//a[text()='Click for treatment schema']"));
-    }
-
-    @Test
-    public void testLearnAboutStudyProducts()
-    {
-        cds.viewLearnAboutPage("Study products");
-        sleep(30000);
-        refresh();
-        sleep(30000);
-        waitForElement(LabKeyPage.Locators.pageSignal("determinationLearnAboutStudyProductLoaded"), 60000);
-        waitForElement(Locator.xpath("//h2").withText("verapamil hydrochloride"), 60000);
-//        longWait().until(ExpectedConditions.visibilityOfElementLocated(Locator.xpath("//div[contains(@class, 'learnview')]//span//div//div[contains(@class, 'learnstudyproducts')]//div[contains(@class, 'learncolumnheader')]").toBy()));
-
-        List<String> studyProducts = Arrays.asList(CDSHelper.PRODUCTS);
-        _asserts.verifyLearnAboutPage(studyProducts);
-    }
-
-    @Test
-    public void clickOnLearnAboutStudyProductsItem()
-    {
-        List<WebElement> returnedItems;
-        String[] itemParts;
-        final String XPATH_RESULTLIST = "//div[contains(@class, 'learnview')]//span//div//div[contains(@class, 'learnstudyproducts')]//div[contains(@class, 'learncolumnheader')]/./following-sibling::div[contains(@class, 'detail-container')]";
-
-        cds.viewLearnAboutPage("Study products");
-        sleep(30000);
-        refresh();
-        sleep(30000);
-        waitForElement(LabKeyPage.Locators.pageSignal("determinationLearnAboutStudyProductLoaded"), 60000);
-        waitForElement(Locator.xpath("//h2").withText("verapamil hydrochloride"), 60000);
-//        longWait().until(ExpectedConditions.visibilityOfElementLocated(Locator.xpath("//div[contains(@class, 'learnview')]//span//div//div[contains(@class, 'learnstudyproducts')]//div[contains(@class, 'learncolumnheader')]").toBy()));
-        returnedItems = Locator.xpath(XPATH_RESULTLIST).findElements(getDriver());
-
-        int index = returnedItems.size()/2;
-
-        itemParts = returnedItems.get(index).getText().split("\n");
-        log("Looking for product: " + itemParts[0] + " in a list of " + returnedItems.size());
-        longWait().until(ExpectedConditions.visibilityOf(returnedItems.get(index)));
-        returnedItems.get(index).click();
-
-        log("Validating title is " + itemParts[0]);
-        longWait().until(ExpectedConditions.visibilityOfElementLocated(Locator.xpath("//div[contains(@class, 'learnheader')]//div//span[text()='" + itemParts[0] + "']").toBy()));
-
-        log("Validating Product Type is: " + itemParts[1]);
-        assert(Locator.xpath("//table[contains(@class, 'learn-study-info')]//tbody//tr//td[contains(@class, 'item-value')][text()='" + itemParts[1] + "']").findElement(getDriver()).isDisplayed());
-
-        log("Validating Class is: " + itemParts[2]);
-        assert(Locator.xpath("//table[contains(@class, 'learn-study-info')]//tbody//tr//td[contains(@class, 'item-value')][text()='" + itemParts[2] + "']").findElement(getDriver()).isDisplayed());
-
-        // TODO could add more code here to validate other fields, but in the interest of time leaving it at this for now.
-
-        log("Validating return link works.");
-        click(Locator.xpath("//div[contains(@class, 'learn-up')]/span[contains(@class, 'breadcrumb')][text()='Study products / ']"));
-
-        shortWait().until(ExpectedConditions.visibilityOfElementLocated(Locator.xpath("//div[contains(@class, 'title')][text()='Learn about...']").toBy()));
-    }
-
-    @Test
-    public void testLearnAboutStudyProductsSearch()
-    {
-        String searchString;
-        List<WebElement> returnedItems;
-        String itemText;
-        String[] itemParts;
-        final String XPATH_TEXTBOX = "//table[contains(@class, 'learn-search-input')]//tbody//tr//td//input";
-        final String XPATH_RESULTLIST = "//div[contains(@class, 'learnview')]//span//div//div[contains(@class, 'learnstudyproducts')]//div[contains(@class, 'learncolumnheader')]/./following-sibling::div[contains(@class, 'detail-wrapper')]";
-
-        cds.viewLearnAboutPage("Study products");
-
-        searchString = "AID"; // TODO Test data dependent.
-        log("Searching for '" + searchString + "'.");
-        this.setFormElement(Locator.xpath(XPATH_TEXTBOX), searchString);
-        sleep(CDSHelper.CDS_WAIT_ANIMATION);  // Same elements are reused between searched, this sleep prevents a "stale element" error.
-        returnedItems = Locator.xpath(XPATH_RESULTLIST).findElements(getDriver());
-        log("Size: " + returnedItems.size());
-        for (WebElement listItem : returnedItems)
-        {
-            itemText = listItem.getText();
-            itemParts = itemText.split("\n");
-            log("Looking at study " + itemParts[0]);
-            assert(itemText.toLowerCase().contains(searchString.toLowerCase()));
-        }
-
-        searchString = "inhibitor"; // TODO Test data dependent.
-        log("Searching for '" + searchString + "'.");
-        this.setFormElement(Locator.xpath(XPATH_TEXTBOX), searchString);
-        sleep(CDSHelper.CDS_WAIT_ANIMATION);
-        returnedItems = Locator.xpath(XPATH_RESULTLIST).findElements(getDriver());
-        for (WebElement listItem : returnedItems)
-        {
-            itemText = listItem.getText();
-            itemParts = itemText.split("\n");
-            log("Looking at study " + itemParts[0]);
-            assert(itemText.toLowerCase().contains(searchString.toLowerCase()));
-        }
-
-        searchString = "GSK"; // TODO Test data dependent.
-        log("Searching for '" + searchString + "'.");
-        this.setFormElement(Locator.xpath(XPATH_TEXTBOX), searchString);
-        sleep(CDSHelper.CDS_WAIT_ANIMATION);
-        returnedItems = Locator.xpath(XPATH_RESULTLIST).findElements(getDriver());
-        for (WebElement listItem : returnedItems)
-        {
-            itemText = listItem.getText();
-            itemParts = itemText.split("\n");
-            log("Looking at study " + itemParts[0]);
-            assert(itemText.toLowerCase().contains(searchString.toLowerCase()));
-        }
-
-        searchString = "is a"; // TODO Test data dependent.
-        log("Searching for '" + searchString + "'.");
-        this.setFormElement(Locator.xpath(XPATH_TEXTBOX), searchString);
-        sleep(CDSHelper.CDS_WAIT_ANIMATION);
-        returnedItems = Locator.xpath(XPATH_RESULTLIST).findElements(getDriver());
-        for (WebElement listItem : returnedItems)
-        {
-            itemText = listItem.getText();
-            itemParts = itemText.split("\n");
-            log("Looking at study " + itemParts[0]);
-            assert(itemText.toLowerCase().contains(searchString.toLowerCase()));
-        }
-
-        searchString = "If this string ever appears something very odd happened.";
-        log("Searching for '" + searchString + "'.");
-        sleep(CDSHelper.CDS_WAIT_ANIMATION);
-        this.setFormElement(Locator.xpath(XPATH_TEXTBOX), searchString);
-        _asserts.verifyEmptyLearnAboutStudyProductsPage();
-
-    }
-
-    @Test
-    public void testLearnAboutAssays()
-    {
-        cds.viewLearnAboutPage("Assays");
-        List<String> assays = Arrays.asList(CDSHelper.ASSAYS_FULL_TITLES);
-        _asserts.verifyLearnAboutPage(assays); // Until the data is stable don't count the assay's shown.
-
-        waitAndClick(Locator.tagWithClass("div", "detail-container").append("/div/div/h2").containing(assays.get(0)));
-        waitForElement(Locator.tagWithClass("span", "breadcrumb").containing("Assays /"));
-        assertTextPresent(CDSHelper.LEARN_ABOUT_BAMA_ANALYTE_DATA);
-
-        //testing variables page
-        waitAndClick(Locator.tagWithClass("h1", "lhdv").withText("Variables"));
-        waitForElement(Locator.tagWithClass("div", "list-entry-container"));
-        assertTextPresent(CDSHelper.LEARN_ABOUT_BAMA_VARIABLES_DATA);
-
-        refresh();
-
-        waitForElement(Locator.tagWithClass("div", "list-entry-container"));
-        assertTextPresent(CDSHelper.LEARN_ABOUT_BAMA_VARIABLES_DATA);
-
-        //testing BAMA antigens page
-        waitAndClick(Locator.tagWithClass("h1", "lhdv").withText("Antigens"));
-        waitForElement(Locator.tagWithClass("div", "list-title-bar").append("/div").containing("Antigen"));
-        assertTextPresent(CDSHelper.LEARN_ABOUT_BAMA_ANTIGEN_DATA);
-
-        refresh(); //refreshes are necessary to clear previously viewed tabs from the DOM.
-
-        //testing ICS antigens page
-        waitAndClick(Locator.tagWithClass("span", "breadcrumb").containing("Assays /"));
-        waitAndClick(Locator.tagWithClass("div", "detail-container").append("/div/div/h2").containing(assays.get(1)));
-        waitForElement(Locator.tagWithClass("span", "breadcrumb").containing("Assays /"));
-
-        refresh();
-
-        waitAndClick(Locator.tagWithClass("h1", "lhdv").withText("Antigens"));
-        waitForElement(Locator.tagWithClass("div", "list-title-bar").append("/div").containing("Protein Panel"));
-        waitForText(CDSHelper.LEARN_ABOUT_ICS_ANTIGEN_TAB_DATA[0]);
-        assertTextPresent(CDSHelper.LEARN_ABOUT_ICS_ANTIGEN_TAB_DATA);
-    }
-
-    @Test
     public void testSummaryPageSingleAxisLinks()
     {
         Locator hierarchySelector = Locator.input("sae-hierarchy");
